Fixing cache

+++ system/library/cache.php
@@ -2,32 +2,27 @@
 class Cache { 
 	private $expire = 3600; 
 
-	public function get($key) {
+	 public function get($key) {
+		$data = null;
 		$files = glob(DIR_CACHE . 'cache.' . preg_replace('/[^A-Z0-9\._-]/i', '', $key) . '.*');
-
 		if ($files) {
-			$cache = file_get_contents($files[0]);
-			
-			$data = unserialize($cache);
-			
-			foreach ($files as $file) {
-				$time = substr(strrchr($file, '.'), 1);
-
-      			if ($time < time()) {
-					if (file_exists($file)) {
-						unlink($file);
-					}
-      			}
-    		}
-			
-			return $data;			
-		}
+		for ($n=0, $lenght = count($files); $n < $lenght; $n++) {
+		$file = $files[$n];
+		$time = substr(strrchr($file, '.'), 1);
+		      if ($time < time()) {
+			if (file_exists($file)) { unlink($file); } 
+		      } elseif (!$n) { 
+			$cache = file_get_contents($file); $data = unserialize($cache);
+			} 
+		} 
+		}  
+		return $data; 
 	}
 
-  	public function set($key, $value) {
+  	public function set($key, $value, $expire=false) {
     	$this->delete($key);
-		
-		$file = DIR_CACHE . 'cache.' . preg_replace('/[^A-Z0-9\._-]/i', '', $key) . '.' . (time() + $this->expire);
+		if (!$expire) $expire=$this->expire;
+		$file = DIR_CACHE . 'cache.' . preg_replace('/[^A-Z0-9\._-]/i', '', $key) . '.' . (time() + $expire);
     	
 		$handle = fopen($file, 'w');
